backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
gunicorn
packaging
PyJWT
six
sqlparse
tzdatas
